refactor: reorganize code for extension tests
